Move around pytest fixtures
