+++ tests/conftest.py
@@ -6,16 +6,6 @@ import pytest
 if TYPE_CHECKING:
     from pathlib import Path
 
-    from typer.testing import CliRunner
-
-
-@pytest.fixture(scope="session")
-def cli() -> "CliRunner":
-    """Fixture for CLI runner."""
-    from typer.testing import CliRunner
-
-    return CliRunner(mix_stderr=False)
-
 
 @pytest.fixture(scope="session")
 def samples() -> "Path":
